requirements.txt
# Instagram Automation GUI - Local Dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
Pillow>=10.0.0

# Optional but recommended
beautifulsoup4>=4.12.0 
# Scheduling and system tray functionality
schedule>=1.2.0
pystray>=0.19.4
